api/requirements.txt
# FastAPI core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async HTTP client
httpx>=0.26.0

# File handling
aiofiles>=23.2.1
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Caching
cachetools>=5.3.0
google-cloud-storage>=2.14.0

# Production server
gunicorn>=21.2.0
